+++ hlrad/lightmap.cpp
@@ -153,7 +153,7 @@ void AddFaceForVertexNormal_printerror (const int edgeabs, const int edgeend, df
 			Log (" e=%d v0=%d(%f,%f,%f) v1=%d(%f,%f,%f) share0=%d share1=%d\n", e,
 				v0, g_dvertexes[v0].point[0], g_dvertexes[v0].point[1], g_dvertexes[v0].point[2],
 				v1, g_dvertexes[v1].point[0], g_dvertexes[v1].point[1], g_dvertexes[v1].point[2],
-				(es->faces[0]==NULL? -1: es->faces[0]-g_dfaces), (es->faces[1]==NULL? -1: es->faces[1]-g_dfaces));
+				(es->faces[0]==NULL? -1: (int)(es->faces[0]-g_dfaces)), (es->faces[1]==NULL? -1: (int)(es->faces[1]-g_dfaces)));
 		}
 	}
 }
@@ -821,7 +821,7 @@ static void     CalcFaceExtents(lightinfo_t* l)
 			)
 		{
 			ThreadLock();
-			PrintOnce("\nfor Face %d (texture %s) at ", s - g_dfaces, TextureNameFromFace(s));
+			PrintOnce("\nfor Face %d (texture %s) at ", (int)(s - g_dfaces), TextureNameFromFace(s));
 
 			for (i = 0; i < s->numedges; i++)
 			{
